--- lib/engine/src/detail/dear_imgui.cpp
@@ -1,111 +1,0 @@
-#include <backends/imgui_impl_glfw.h>
-#include <backends/imgui_impl_vulkan.h>
-#include <imgui.h>
-#include <detail/dear_imgui.hpp>
-#include <facade/util/logger.hpp>
-#include <facade/vk/cmd.hpp>
-#include <glm/gtc/color_space.hpp>
-#include <glm/mat4x4.hpp>
-
-namespace facade {
-namespace {
-vk::UniqueDescriptorPool make_pool(vk::Device const device) {
-	vk::DescriptorPoolSize pool_sizes[] = {
-		{vk::DescriptorType::eSampledImage, 1000},		   {vk::DescriptorType::eCombinedImageSampler, 1000}, {vk::DescriptorType::eSampledImage, 1000},
-		{vk::DescriptorType::eStorageImage, 1000},		   {vk::DescriptorType::eUniformTexelBuffer, 1000},	  {vk::DescriptorType::eStorageTexelBuffer, 1000},
-		{vk::DescriptorType::eUniformBuffer, 1000},		   {vk::DescriptorType::eStorageBuffer, 1000},		  {vk::DescriptorType::eUniformBufferDynamic, 1000},
-		{vk::DescriptorType::eStorageBufferDynamic, 1000}, {vk::DescriptorType::eInputAttachment, 1000},
-	};
-	auto pool_info = vk::DescriptorPoolCreateInfo{};
-	pool_info.flags = vk::DescriptorPoolCreateFlagBits::eFreeDescriptorSet;
-	pool_info.maxSets = 1000 * std::size(pool_sizes);
-	pool_info.poolSizeCount = static_cast<std::uint32_t>(std::size(pool_sizes));
-	pool_info.pPoolSizes = pool_sizes;
-	return device.createDescriptorPoolUnique(pool_info);
-}
-
-glm::vec4 to_linear(glm::vec4 const& srgb) { return glm::convertSRGBToLinear(srgb); }
-
-void correct_style() {
-	auto* colours = ImGui::GetStyle().Colors;
-	for (int i = 0; i < ImGuiCol_COUNT; ++i) {
-		auto& colour = colours[i];
-		auto const corrected = to_linear(glm::vec4{colour.x, colour.y, colour.z, colour.w});
-		colour = {corrected.x, corrected.y, corrected.z, corrected.w};
-	}
-}
-
-DearImGui const* g_instance{};
-} // namespace
-
-std::unique_ptr<DearImGui> DearImGui::make(CreateInfo const& create_info) {
-	if (g_instance) {
-		logger::error("Attempt to create duplicate Dear ImGui instance denied");
-		return {};
-	}
-
-	auto pool = make_pool(create_info.gfx.device);
-
-	IMGUI_CHECKVERSION();
-	ImGui::CreateContext();
-	ImGuiIO& io = ImGui::GetIO();
-	io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
-	io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;  // Enable Gamepad Controls
-
-	ImGui::StyleColorsDark();
-	if (create_info.swapchain == ColourSpace::eSrgb) { correct_style(); }
-
-	auto loader = vk::DynamicLoader{};
-	auto get_fn = [&loader](char const* name) { return loader.getProcAddress<PFN_vkVoidFunction>(name); };
-	auto lambda = +[](char const* name, void* ud) {
-		auto const* gf = reinterpret_cast<decltype(get_fn)*>(ud);
-		return (*gf)(name);
-	};
-	ImGui_ImplVulkan_LoadFunctions(lambda, &get_fn);
-	ImGui_ImplGlfw_InitForVulkan(create_info.window, true);
-	ImGui_ImplVulkan_InitInfo init_info = {};
-	init_info.Instance = create_info.gfx.instance;
-	init_info.PhysicalDevice = create_info.gfx.gpu;
-	init_info.Device = create_info.gfx.device;
-	init_info.QueueFamily = create_info.gfx.queue_family;
-	init_info.Queue = create_info.gfx.queue;
-	init_info.DescriptorPool = *pool;
-	init_info.Subpass = 0;
-	init_info.MinImageCount = 2;
-	init_info.ImageCount = 2;
-	init_info.MSAASamples = static_cast<VkSampleCountFlagBits>(create_info.samples);
-
-	ImGui_ImplVulkan_Init(&init_info, create_info.render_pass);
-
-	{
-		auto cmd = Cmd{create_info.gfx};
-		ImGui_ImplVulkan_CreateFontsTexture(cmd.cb);
-	}
-	ImGui_ImplVulkan_DestroyFontUploadObjects();
-
-	return std::unique_ptr<DearImGui>(new DearImGui{std::move(pool)});
-}
-
-DearImGui::DearImGui(vk::UniqueDescriptorPool pool) : m_pool(std::move(pool)) { g_instance = this; }
-
-DearImGui::~DearImGui() {
-	m_pool.getOwner().waitIdle();
-	ImGui_ImplVulkan_Shutdown();
-	ImGui_ImplGlfw_Shutdown();
-	ImGui::DestroyContext();
-	g_instance = {};
-}
-
-void DearImGui::new_frame() {
-	ImGui_ImplVulkan_NewFrame();
-	ImGui_ImplGlfw_NewFrame();
-	ImGui::NewFrame();
-}
-
-void DearImGui::end_frame() {
-	// ImGui::Render calls ImGui::EndFrame
-	ImGui::Render();
-}
-
-void DearImGui::render(vk::CommandBuffer const cb) { ImGui_ImplVulkan_RenderDrawData(ImGui::GetDrawData(), cb); }
-} // namespace facade
